+++ NEWS.md
@@ -1,6 +1,14 @@
 1.16.1 - 2014-mm-dd
 -------------------
 
+Enhancements:
+
+ * Upgrades dependencies:
+    * cartodb-redis
+    * cartodb-psql
+    * log4js
+
+
 
 1.16.0 - 2014-08-19
 -------------------